implement pmtiles extract [#31, #52] (#62)
* implement pmtiles extract [#31, #52]

* Experimental cli support for extracting a region from a larger archive, given a maxzoom and GeoJSON multipolygon region.
* Limited to credentialed buckets or local files now, public HTTP to come later
* Limited to a single download thread
* Change directory optimization to be faster and match Java implementation, affects root/leaf sizes

* Finish initial extract [#31]

* include the DstOffset so we can multithread downloads later
* set header statistics
* implement --dry-run
* add logging messages for user feedback

+++ pmtiles/convert.go
@@ -114,14 +114,14 @@ func Convert(logger *log.Logger, input string, output string, deduplicate bool, 
 func add_directoryv2_entries(dir DirectoryV2, entries *[]EntryV3, f *os.File) {
 	for zxy, rng := range dir.Entries {
 		tile_id := ZxyToId(zxy.Z, zxy.X, zxy.Y)
-		*entries = append(*entries, EntryV3{tile_id, rng.Offset, rng.Length, 1})
+		*entries = append(*entries, EntryV3{tile_id, rng.Offset, uint32(rng.Length), 1})
 	}
 
 	var unique = map[uint64]uint32{}
 
 	// uniqify the offset/length pairs
 	for _, rng := range dir.Leaves {
-		unique[rng.Offset] = rng.Length
+		unique[rng.Offset] = uint32(rng.Length)
 	}
 
 	for offset, length := range unique {
